refactor : 도메인 - 인프라 연결 후 삭제될 파일 정리
