+++ src/main/java/com/icestudyroom_email/domain/common/kafka/producer/KafkaRankingEventPublisher.java
@@ -1,7 +1,7 @@
 package com.icestudyroom_email.domain.common.kafka.producer;
 
-import com.icestudyroom_email.domain.rankingContract.RankingEmailEvent;
-import com.icestudyroom_email.domain.rankingContract.forDelete.RankingEmailEventPublisher;
+import com.icestudyroom_email.domain.contract.ranking.RankingEmailEvent;
+import com.icestudyroom_email.domain.rankingContract.forTest.RankingEmailEventPublisher;
 import lombok.RequiredArgsConstructor;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Component;
