+++ src/main/java/com/icestudyroom_email/domain/rankingContract/forTest/TestController.java
@@ -1,6 +1,6 @@
-package com.icestudyroom_email.domain.rankingContract.forDelete;
+package com.icestudyroom_email.domain.rankingContract.forTest;
 
-import com.icestudyroom_email.domain.rankingContract.WeeklyRankingDto;
+import com.icestudyroom_email.domain.contract.ranking.WeeklyRankingDto;
 import com.icestudyroom_email.domain.socket.infrastructure.protocol.ranking.RankingSocketEvent;
 import com.icestudyroom_email.domain.socket.infrastructure.ranking.SocketRankingBroadcaster;
 import lombok.RequiredArgsConstructor;
